+++ tools/loadtest/ui/src/main/java/is/codion/tools/loadtest/ui/LoadTestPanel.java
@@ -553,8 +553,9 @@ public final class LoadTestPanel<T> extends JPanel {
 
 		@Override
 		public FilterTableCellRenderer tableCellRenderer(FilterTableColumn<ColumnId> column) {
-			FilterTableCellRenderer.Builder<ApplicationRow, ColumnId> builder =
-							FilterTableCellRenderer.builder(model().applicationTableModel(), column.identifier(), Integer.class);
+			FilterTableCellRenderer.Builder<ColumnId> builder =
+							FilterTableCellRenderer.builder(column.identifier(), Integer.class)
+											.condition(model().applicationTableModel().conditions().optional(column.identifier()).orElse(null));
 			if (column.identifier().equals(ColumnId.DURATION)) {
 				builder.string(duration -> duration == null ? null : DURATION_FORMAT.format(duration));
 			}